+++ NEWS.md
@@ -4,6 +4,9 @@
 * Prototype `Builder` for building Debian packages (#47)
 * Added option `--bind-volatile`, working as `--bind-ro` plus a temporary
   writable overlay (#50)
+* Added option `--workdir-volatile` working as `--workdir` but with a volatile
+  mount (#50)
+* Use volatile mounts for CI (#51)
 
 # Version 0.6